tests/wpt/css-tests/cssom-view-1_dev/xhtml1/window-screen-width.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <title>CSSOM View - 4.2 - screen.width range tests</title>
    <link href="mailto:neils.christoffersen@gmail.com" rel="author" title="Neils Christoffersen" />
	<link href="http://www.w3.org/TR/cssom-view/#the-screen-interface" rel="help" />
    <meta content="dom" name="flags" />
    <meta content="window.screen.width has sensible values" name="assert" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <div id="myDiv"></div>
    <div id="log"></div>
    <script>
		var upperBound = 50000;

        test(
			function() {
				assert_not_equals(window.screen.width, 0, "window.screen.width is zero")
			}, "zero check"
		);

		test(
			function() {
				assert_true(window.screen.width &gt; 0, "window.screen.width shouldn't be negative")
			}, "positive check"
		);

		test(
			function() {
				assert_true(window.screen.width &lt; upperBound, "window.screen.width shouldn't be so large")
			}, "upper bound check"
		);
    </script>


</body></html>
